refactor: replace room functions
